requirements.txt
quart==0.20.0
gunicorn==21.2.0
uvicorn==0.29.0
aiohttp==3.9.5
google-generativeai==0.7.2
requests==2.31.0
pandas==2.2.1
python-dotenv==1.0.1
numpy==1.26.4
